web interface: remove note about misleading mtime of python scripts.
mtime is modification date and will not match the release date of the
actual python code running.  It shows the install date, which may even
be yesterday for a version of ParticleStats that is 10 years old.  This
makes it only useful if ParticleStats was installed on day of release.
Remove misleading information from HTML output. If we want something
like this, then the scripts should check the version of the ParticleStats
pacage (which is available via pkg_resources for example).