Reduce global variables, remove unused variables, and limit scope of variables.
Global variables which are then locally masked make things hard to
understand.  So does a lot of code evaluation in the global scope.
Enclose things in sub and only pass required variables around.  Also
define variables only when needed and only on the scope they are
needed.

Check again on Process sub if there it was choosen a valid step.